virtinstall: avoid duplicating output on Xen guest reboot
This patch is addressing duplicate output in RHEL5 xen guest console
logs. Problem is that when xen guest restarts, "virsh console" process
will end. zrhel5_write_consolelogs.py will restart it, but output
is getting appened to console log file.
logguestconsoles.c will notice this via inotify IN_CLOSE_WRITE and
resets file offset to begining, which is wrong, because file wasn't truncated.

This patch keeps track of latest offset and compares it to file size.
If latest offset is higher than file size it means that file was
truncated and offset should be reset to begining of file.

Signed-off-by: Jan Stancek <jstancek@redhat.com>

Bug: 1048776
Change-Id: I49bd2f52d1e2001372d5021bf3b3923a6faafe48