virt/install: adding support for macvtap interfaces [v2]
The --network portion of the guestargs are overwritten by runtest.sh and
are changed from:

--network type=direct,source=em1,source_mode=bridge,target=macvtap1

to

--network bridge:br3,type=direct,source=em1,source_mode=bridge,target=macvtap1

This syntax is invalid for virt-install *and* there appears to be no way
to enable macvtap-based interfaces at all when using the beaker virt
install test.  The modified line above is invalid for virt-install.

This patch checks to see if 'type=direct' is included and does not add
the optional 'bridge:...' entry to any network arguments.

v2: Simplified parsing/matching of network arguments as suggeted by Dan
Callaghan <dcallagh@redhat.com>

Bug: 1086285

Change-Id: I3ff3364c9ca08a8bb5c134275e775b19598c0a3d