virt/install: don't submit libvirtd debug logs by default
In recent versions of libvirt the debug log can be massive. So we will
continue to configure libvirtd to write the debug logs to disk (in
/var/tmp rather than /tmp to avoid filling up a tmpfs) however we will
no longer upload the log to Beaker by default.

Setting the LIBVIRTD_DEBUG task param to a non-empty value will cause
the log to be uploaded to Beaker as before.

Bug: 1117001
Change-Id: Iad11cbea79ceae5953d4a53ee0961d2838f6ebf7