virt: set expected MACADDR on bridge interface
The scripts assume that the bridge will reuse the same MACADDR as the
network interface, so that the system will still be accessible through
the bridge under the same address.
This is not enforced and may not hold true in some systems, so specify
the MACADDR in the bridge interface.

Signed-off-by: Erico Nunes <ernunes@redhat.com>