virt: move python scripts to versioned subdirectory
This is in preparation to enable python3 support.
Move the python scripts to a subdirectory, and add some logic to copy
them over so everything still works.
This commit mostly moves the scripts without any modifications.
In a followup commit, we can add replacement python3 scripts for newer
systems where python2 may not be available.

Signed-off-by: Erico Nunes <ernunes@redhat.com>